requirements.txt
# Core dependencies
openai==4.0.0
pinecone-client==2.0.1
rank-bm25==0.2.1
langchain==0.0.158
streamlit==1.17.0
python-dotenv==1.0.0
scrapy==2.5.1
pandas==1.3.3
numpy==1.21.2